+++ src/test/groovy/me/xdrop/jrand/generators/basics/DoubleGeneratorTest.groovy
@@ -4,11 +4,11 @@ class DoubleGeneratorTest extends GroovyTestCase {
     def instance = {-> new DoubleGenerator()}
 
     void testMax() {
-        assertTrue instance().max(2.0).gen() <= 2
+        assertTrue instance().max(2.0 as double).gen() <= 2
     }
 
     void testMin() {
-        assertTrue instance().min(5.0).gen() >= 5
+        assertTrue instance().min(5.0 as double).gen() >= 5
     }
 
     void testRange() {
